# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Banking API",
    "version": "v1.0.0"
  },
  "paths": {
    "/transaction/withdraw/{accountId}": {
      "post": {
        "summary": "Withdraw money",
        "parameters": [
          { "name": "accountId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/TransactionRequest" } }
          }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object" }, "example": { "status": "success" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
          }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object" }, "example": { "status": "created" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object" }, "example": { "token": "abc-123" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/expense/summary/{accountId}": {
      "get": {
        "summary": "Get budget summary",
        "parameters": [
          { "name": "accountId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object" }, "example": { "total": 500.0 } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TransactionRequest": { "type": "object", "required": ["amount"], "properties": { "amount": { "type": "number" } } },
      "RegisterRequest": { "type": "object", "required": ["name", "email", "password"], "properties": { "name": { "type": "string" }, "email": { "type": "string" }, "password": { "type": "string" } } },
      "LoginRequest": { "type": "object", "required": ["email", "password"], "properties": { "email": { "type": "string" }, "password": { "type": "string" } } },
      "Error": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "status": { "type": "integer" },
          "error": { "type": "string" },
          "message": { "type": "string" },
          "path": { "type": "string" }
        }
      }
    }
  }
}